+++ src/pygobbler/upload_directory.py
@@ -1,9 +1,10 @@
+from typing import Optional
 import os
 from . import allocate_upload_directory
 from . import _utils as ut
 
 
-def upload_directory(project: str, asset: str, version: str, directory: str, staging: str, url: str, probation: bool = False):
+def upload_directory(project: str, asset: str, version: str, directory: str, staging: str, url: str, probation: bool = False, consume: Optional[bool] = None, ignore_dot: bool = True):
     """
     Upload a directory as a new versioned asset of a project in the registry.
 
@@ -30,11 +31,28 @@ def upload_directory(project: str, asset: str, version: str, directory: str, sta
 
         probation:
             Whether to upload a probational version.
+    
+        consume:
+            Whether the contents of ``directory`` can be consumed by the upload process.
+            If true, the Gobbler will attempt to move files from ``directory`` into the registry.
+            Otherwise, the contents of ``directory`` will not be modified by the upload.
+            Defaults to true if the contents of ``directory`` need to be copied to ``staging``.
+
+        ignore_dot:
+            Whether to skip dotfiles in ``directory`` during upload.
     """
     directory = os.path.normpath(directory)
     staging = os.path.normpath(staging)
 
-    if os.path.dirname(directory) != staging:
+    in_staging = False
+    tmpd = directory
+    while len(tmpd) > len(staging):
+        tmpd = os.path.dirname(tmpd)
+        if tmpd == staging:
+            in_staging = True
+            break
+
+    if not in_staging:
         newdir = allocate_upload_directory(staging) 
 
         for root, dirs, files in os.walk(directory):
@@ -58,12 +76,17 @@ def upload_directory(project: str, asset: str, version: str, directory: str, sta
 
         directory = newdir
 
+    if consume is None:
+        consume = not in_staging
+
     req = {
         "source": os.path.basename(directory),
         "project": project,
         "asset": asset,
         "version": version,
-        "on_probation": probation
+        "on_probation": probation,
+        "consume": consume,
+        "ignore_dot": ignore_dot
     }
     ut.dump_request(staging, url, "upload", req)
     return